Update Project to use Gradle
Update and reformat the structure of the project to use Gradle, simplify
the `src` directory to `com.sonneveld.ToyRobot` package, have Unit tests
running and build the project as `jar` with a Manifest.

Also, add the Toy Robot specification to the project as part of its
rebuild / reformat to start development again.

== Notes:
- Modify all binstubs to use Gradle where appropriate

== References:
- [Gradle Build a JAVA Project]
  (https://www.tutorialspoint.com/gradle/gradle_build_a_java_project.htm)

- [How to create Java gradle project \- Stack Overflow]
  (https://stackoverflow.com/questions/14017364/how-to-create-java-gradle-project)

- [Java Quickstart \- Gradle User Manual]
  (https://docs.gradle.org/current/userguide/tutorial_java_projects.html)

- [Getting Started · Building Java Projects with Gradle]
  (https://spring.io/guides/gs/gradle/)