Add `exec` method to Robot class
Add the execute method `exec` to the Robot class that accepts two
arguments, a command name as a string and its arguments that is also a
string. This allows for outside input into the Robot instance as strings
in order to get call the Robot methods indirectly from the console
captured User input, which will be later used to fulfil the Toy Robot
specification.

== References:
- [switch statement in java]
  (https://www.tutorialspoint.com/java/switch_statement_in_java.htm)

- [Strings in switch Statements]
  (https://docs.oracle.com/javase/7/docs/technotes/guides/language/strings-switch.html)