Add CLI class as entrypoint to main App
Add the Toy Robot CLI class with the `run` and `process` methods for use
within main Application. The `run` method accepts the expect command
line string array of arguments that quickly inspected for a contents
length greater than 0 and then branches to use either the first argument
as a filename for a file to be `process`ed or start accepting User
input.

The `process` method then takes the incoming stream and scans it line by
line until either the end of file is reached or "EXIT" command is given,
otherwise splits each line on a space character and feeds those words as
the command & its command arguments to the Robot instance.

= Notes:
- Update the `bin/run` script to accept the first argument passed to it
  and then pass it directly to the Java `jar` file.

- Also, update the `run` binstub to allow build the `jar` executable
  when it does not exist`

== References:
- [Java 8 Stream – Read a file line by line – Mkyong\.com]
  (https://www.mkyong.com/java8/java-8-stream-read-a-file-line-by-line/)

- [How do I compare strings in Java? \- Stack Overflow]
  (https://stackoverflow.com/questions/513832/how-do-i-compare-strings-in-java)

- [How do I tell if a regular file does not exist in Bash? \- Stack
  Overflow]
  (https://stackoverflow.com/questions/638975/how-do-i-tell-if-a-regular-file-does-not-exist-in-bash)

- [String array initialization in Java \- Stack Overflow]
  (https://stackoverflow.com/questions/17515096/string-array-initialization-in-java)

- [java \- Is it possible to take a string and put it in the System\.in?
  \- Stack Overflow]
  (https://stackoverflow.com/questions/10759399/is-it-possible-to-take-a-string-and-put-it-in-the-system-in)